+++ src/api/HttpLink.js
@@ -1,11 +1,10 @@
 import * as Sentry from '@sentry/vue';
 import { BatchHttpLink } from 'apollo-link-batch-http';
-import fetch from 'isomorphic-fetch';
 // http is a nodejs dep only used during development mode (there is no http npm package as of July 2020)
 import { Agent as HttpAgent } from 'http'; // eslint-disable-line import/no-extraneous-dependencies
 import { Agent as SslAgent } from 'https';
 
-export default ({ kvAuth0, uri = '' }) => {
+export default ({ kvAuth0, uri = '', fetch }) => {
 	const onVm = uri.indexOf('vm') > -1;
 	const usingLocal = uri.indexOf('localhost') > -1;
 	const secure = uri.indexOf('https') === 0;