+++ src/pages/Lend/LoanSearchPage.vue
@@ -1,33 +1,31 @@
 <template>
 	<www-page id="lend-filter-alpha">
 		<article class="tw-bg-secondary tw-relative tw-pt-6">
-			<div>
-				<kv-page-container>
-					<div class="tw-flex tw-items-start tw-pb-8">
-						<div class="tw-flex-1">
-							<h1 class="tw-mb-2">
-								Make a loan, change a life
-							</h1>
-							<p class="tw-text-h3 tw-text-secondary md:tw-w-2/3 tw-mb-3p tw-hidden md:tw-block">
-								Each Kiva loan helps people build a better future for themselves and their families.
-							</p>
-						</div>
-						<button class="tw-mb-2 tw-mt-3 tw-border-r tw-border-tertiary tw-px-1 md:tw-px-2">
-							<kv-material-icon :icon="mdiEarth" class="tw-text-secondary tw-w-3 tw-h-3" />
-							<p class="tw-text-tertiary tw-hidden md:tw-block">
-								Explore
-							</p>
-						</button>
-						<button class="tw-mb-2  tw-mt-3 tw-px-1 md:tw-px-2">
-							<kv-material-icon :icon="mdiFilter" class="tw-text-brand tw-w-3 tw-h-3" />
-							<p class="tw-text-tertiary tw-hidden md:tw-block">
-								Filters
-							</p>
-						</button>
+			<kv-page-container>
+				<div class="tw-flex tw-items-start tw-pb-8">
+					<div class="tw-flex-1">
+						<h1 class="tw-mb-2">
+							Make a loan, change a life
+						</h1>
+						<p class="tw-text-h3 tw-text-secondary md:tw-w-2/3 tw-mb-3p tw-hidden md:tw-block">
+							Each Kiva loan helps people build a better future for themselves and their families.
+						</p>
 					</div>
-					<loan-search-interface />
-				</kv-page-container>
-			</div>
+					<button class="tw-mb-2 tw-mt-3 tw-border-r tw-border-tertiary tw-px-1 md:tw-px-2">
+						<kv-material-icon :icon="mdiEarth" class="tw-text-secondary tw-w-3 tw-h-3" />
+						<p class="tw-text-tertiary tw-hidden md:tw-block">
+							Explore
+						</p>
+					</button>
+					<button class="tw-mb-2  tw-mt-3 tw-px-1 md:tw-px-2">
+						<kv-material-icon :icon="mdiFilter" class="tw-text-brand tw-w-3 tw-h-3" />
+						<p class="tw-text-tertiary tw-hidden md:tw-block">
+							Filters
+						</p>
+					</button>
+				</div>
+				<loan-search-interface />
+			</kv-page-container>
 		</article>
 	</www-page>
 </template>