+++ src/router/routes.js
@@ -49,6 +49,12 @@ module.exports = [
 	{ path: '/build/getting-started', component: () => import('@/pages/Build/GettingStartedPage') },
 	{ path: '/build/research', component: () => import('@/pages/Build/Research') },
 	{ path: '/build/terms-of-service', component: () => import('@/pages/Build/TermsOfService') },
+
+	{
+		path: '/categories-beta',
+		component: () => import('@/pages/Categories/CategoriesBeta')
+	},
+
 	{
 		path: '/cc/:dynamicRoute',
 		component: () => import('@/pages/LandingPages/CorporateCampaign/CCLandingPage'),