+++ src/util/loanSearchUtils.js
@@ -41,6 +41,8 @@ export const sortByNameToDisplay = {
 	personalized: 'Recommended'
 };
 
+export const visibleThemeIds = [2, 6, 8, 11, 14, 28, 29, 32, 36];
+
 /**
  * Returns loan search state that has been validated against the available facets
  *
@@ -129,12 +131,14 @@ export function formatSortOptions(standardSorts, flssSorts) {
 			sortSrc: STANDARD_QUERY_TYPE,
 		};
 	});
-	const labeledFlssSorts = flssSorts.map(sort => {
-		return {
-			name: sort.name,
-			sortSrc: FLSS_QUERY_TYPE,
-		};
-	});
+	const labeledFlssSorts = flssSorts.reduce((prev, current) => {
+		// The amountLeft sort is currently returned by the GraphQL enum but isn't fully supported
+		if (current.name !== 'amountLeft') {
+			prev.push({ name: current.name, sortSrc: FLSS_QUERY_TYPE });
+		}
+
+		return prev;
+	}, []);
 	return [...labeledStandardSorts, ...labeledFlssSorts];
 }
 
@@ -215,11 +219,17 @@ export function transformSectors(filteredSectors, allSectors = []) {
 export function transformThemes(filteredThemes, allThemes = []) {
 	const transformed = [];
 
-	filteredThemes.forEach(({ key: name, value: numLoansFundraising }) => {
+	// Always show certain themes regardless of whether there are applicable loans
+	visibleThemeIds.forEach(id => {
+		const themeFromLend = allThemes.find(a => a.id === id);
+
+		if (!themeFromLend) return;
+
 		// Case insensitive matching since lend and FLSS APIs can use different casing for themes
-		const lookupTheme = allThemes.find(a => a.name.toUpperCase() === name.toUpperCase());
-		if (!lookupTheme) return;
-		const theme = { id: lookupTheme.id, name, numLoansFundraising };
+		const themeFromFlss = filteredThemes.find(t => t.key.toUpperCase() === themeFromLend.name.toUpperCase());
+
+		const theme = { id, name: themeFromLend.name, numLoansFundraising: themeFromFlss?.value ?? 0 };
+
 		transformed.push(theme);
 	});
 
@@ -510,19 +520,21 @@ export function getThemeNamesQueryParam(param, facets) {
  * @param {Object} query The Vue Router query object (this.$route.query)
  * @param {Object} allFacets All available facets from the APIs
  * @param {string} queryType The current query type (lend vs FLSS)
+ * @param {number} pageLimit The limit/size of the page
  * @param {Object} previousState The previous search state
  */
-export async function applyQueryParams(apollo, query, allFacets, queryType, previousState) {
-	// Convert query param 1-based page to pager 0-based page
-	const page = isNumber(query.page) && query.page > 0 ? query.page - 1 : 0;
+export async function applyQueryParams(apollo, query, allFacets, queryType, pageLimit, previousState = {}) {
+	// Convert query param 1-based page to pager 0-based page and ensure page is an integer
+	const page = isNumber(query.page) && query.page >= 1 ? Math.floor(query.page) - 1 : 0;
 
 	const filters = {
-		...previousState, // The countryIsoCode and pageLimit are not currently in the query params
 		gender: query.gender,
-		sortBy: queryType === FLSS_QUERY_TYPE ? lendToFlssSort.get(query.sortBy) : query.sortBy,
+		countryIsoCode: previousState.countryIsoCode,
 		sectorId: getSectorIdsFromQueryParam(query.sector, allFacets.sectorNames, allFacets.sectorFacets, true),
+		sortBy: queryType === FLSS_QUERY_TYPE ? lendToFlssSort.get(query.sortBy) : query.sortBy,
 		theme: getThemeNamesQueryParam(query.attribute, allFacets.themeFacets),
-		pageOffset: page * previousState.pageLimit,
+		pageOffset: page * pageLimit,
+		pageLimit,
 	};
 
 	await updateSearchState(apollo, filters, allFacets, queryType, previousState);